requirements.txt
openai>=0.27.0
fastapi>=0.94.0
uvicorn>=0.21.0
pydantic>=1.10.6
python-dotenv>=1.0.0
langchain>=0.0.113
httpx>=0.23.0
pytest>=7.2.0
google-cloud-secret-manager>=2.16.0